pyproject.toml
[tool.poetry]
name = "the-orchard-check"
version = "0.1.0"
description = ""
authors = ["Andrey Tarakanov <andrey.t-krd@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"
logging = "^0.4.9.6"
oauth2client = "^4.1.3"
google-api-python-client = "^2.121.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
pygsheets = "^2.0.6"
httplib2 = "^0.22.0"
apiclient = "^1.0.4"
loguru = "^0.7.2"
config = "^0.5.1"
pandas = "^2.2.1"
gspread = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
